mdbook_core::config

Struct HtmlConfig

#[non_exhaustive]
pub struct HtmlConfig {

    pub theme: Option<PathBuf>,
    pub default_theme: Option<String>,
    pub preferred_dark_theme: Option<String>,
    pub smart_punctuation: bool,
    pub definition_lists: bool,
    pub admonitions: bool,
    pub mathjax_support: bool,
    pub additional_css: Vec<PathBuf>,
    pub additional_js: Vec<PathBuf>,
    pub fold: Fold,
    pub playground: Playground,
    pub code: Code,
    pub print: Print,
    pub no_section_label: bool,
    pub search: Option<Search>,
    pub git_repository_url: Option<String>,
    pub git_repository_icon: Option<String>,
    pub input_404: Option<String>,
    pub site_url: Option<String>,
    pub cname: Option<String>,
    pub edit_url_template: Option<String>,
    pub redirect: HashMap<String, String>,
    pub hash_files: bool,
    pub sidebar_header_nav: bool,
    /* private fields */

}

Configuration for the HTML renderer.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

theme: Option<PathBuf>

The theme directory, if specified.

default_theme: Option<String>

The default theme to use, defaults to โlightโ

preferred_dark_theme: Option<String>

The theme to use if the browser requests the dark version of the site. Defaults to โnavyโ.

smart_punctuation: bool

Supports smart quotes, apostrophes, ellipsis, en-dash, and em-dash.

definition_lists: bool

Support for definition lists.

admonitions: bool

Support for admonitions.

mathjax_support: bool

Should mathjax be enabled?

additional_css: Vec<PathBuf>

Additional CSS stylesheets to include in the rendered pageโs <head>.

additional_js: Vec<PathBuf>

Additional JS scripts to include at the bottom of the rendered pageโs <body>.

fold: Fold

Fold settings.

playground: Playground

Playground settings.

code: Code

Code settings.

print: Print

Print settings.

no_section_label: bool

Donโt render section labels.

search: Option<Search>

Search settings. If None, the default will be used.

git_repository_url: Option<String>

Git repository url. If None, the git button will not be shown.

git_repository_icon: Option<String>

FontAwesome icon class to use for the Git repository link. Defaults to fa-github if None.

input_404: Option<String>

Input path for the 404 file, defaults to 404.md, set to โโ to disable 404 file output

site_url: Option<String>

Absolute url to site, used to emit correct paths for the 404 page, which might be accessed in a deeply nested directory

cname: Option<String>

The DNS subdomain or apex domain at which your book will be hosted. This string will be written to a file named CNAME in the root of your site, as required by GitHub Pages (see Managing a custom domain for your GitHub Pages site).

edit_url_template: Option<String>

Edit url template, when set shows a โSuggest an editโ button for directly jumping to editing the currently viewed page. Contains {path} that is replaced with chapter source file path

redirect: HashMap<String, String>

The mapping from old pages to new pages/URLs to use when generating redirects.

hash_files: bool

If this option is turned on, โcache bustโ static files by adding hashes to their file names.

The default is true.

sidebar_header_nav: bool

If enabled, the sidebar includes navigation for headers on the current page. Default is true.

Implementations

impl HtmlConfig

pub fn theme_dir(&self, root: &Path) -> PathBuf

Returns the directory of theme from the provided root directory. If the directory is not present it will append the default directory of โthemeโ

pub fn get_404_output_file(&self) -> String

Returns the name of the file used for HTTP 404 โnot foundโ with the .html extension.