mdbook_core::utils

Module fs

Filesystem utilities and helpers.

Functions

copy_files_except_ext

Copies all files of a directory to another one except the files with the extensions given in the ext_blacklist array

create_dir_all

Equivalent to std::fs::create_dir_all with better error messages.

path_to_root

Takes a path and returns a path containing just enough ../ to point to the root of the given path.

read_to_string

Reads a file into a string.

remove_dir_content

Removes all the content of a directory but not the directory itself.

write

Writes a file to disk.